refactor(sandwich-victim): modular detector architecture
