+++ crates/sandwich-victim/src/lib.rs
@@ -9,3 +9,4 @@ pub mod simulation;
 pub mod dex;
 pub mod core;
 pub mod filters;
+pub mod detectors;
